refactor(content-wrapper): rename `content-gutter` to `content-wrapper`
